refactor: SortConstant 리팩터링
Stringvalue를 value로 rename
패키지 명 rename

Related to : #106
