+++ backend/src/main/java/com/prograngers/backend/repository/problem/QueryDslProblemRepository.java
@@ -3,7 +3,7 @@ package com.prograngers.backend.repository.problem;
 import com.prograngers.backend.entity.problem.Problem;
 import com.prograngers.backend.entity.solution.AlgorithmConstant;
 import com.prograngers.backend.entity.solution.DataStructureConstant;
-import com.prograngers.backend.entity.constants.SortConstant;
+import com.prograngers.backend.entity.sortconstant.SortConstant;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
