+++ backend/src/main/java/com/prograngers/backend/repository/problem/QueryDslProblemRepositoryImpl.java
@@ -1,11 +1,9 @@
 package com.prograngers.backend.repository.problem;
 
 import com.prograngers.backend.entity.problem.Problem;
-import com.prograngers.backend.entity.problem.QProblem;
 import com.prograngers.backend.entity.solution.AlgorithmConstant;
 import com.prograngers.backend.entity.solution.DataStructureConstant;
-import com.prograngers.backend.entity.constants.SortConstant;
-import com.prograngers.backend.entity.solution.QSolution;
+import com.prograngers.backend.entity.sortconstant.SortConstant;
 import com.prograngers.backend.exception.enumtype.SortTypeNotFoundException;
 import com.querydsl.core.types.OrderSpecifier;
 import com.querydsl.core.types.dsl.BooleanExpression;
@@ -18,7 +16,7 @@ import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
-import static com.prograngers.backend.entity.constants.SortConstant.*;
+import static com.prograngers.backend.entity.sortconstant.SortConstant.*;
 import static com.prograngers.backend.entity.problem.QProblem.*;
 import static com.prograngers.backend.entity.solution.QSolution.*;
