+++ backend/src/main/java/com/prograngers/backend/repository/solution/QueryDslSolutionRepository.java
@@ -4,7 +4,7 @@ import com.prograngers.backend.entity.solution.Solution;
 import com.prograngers.backend.entity.solution.AlgorithmConstant;
 import com.prograngers.backend.entity.solution.DataStructureConstant;
 import com.prograngers.backend.entity.solution.LanguageConstant;
-import com.prograngers.backend.entity.constants.SortConstant;
+import com.prograngers.backend.entity.sortconstant.SortConstant;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
