+++ backend/src/main/java/com/prograngers/backend/service/ProblemService.java
@@ -6,7 +6,7 @@ import com.prograngers.backend.entity.problem.Problem;
 import com.prograngers.backend.entity.solution.Solution;
 import com.prograngers.backend.entity.solution.AlgorithmConstant;
 import com.prograngers.backend.entity.solution.DataStructureConstant;
-import com.prograngers.backend.entity.constants.SortConstant;
+import com.prograngers.backend.entity.sortconstant.SortConstant;
 import com.prograngers.backend.repository.problem.ProblemRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
